maps and rgb working
