+++ REQUIRE
@@ -20,4 +20,3 @@ Showoff
 Suppressor
 FixedPointNumbers
 Plots
-Measures
